Fix Mixin incompatibility with MixinPlugin
