+++ src/main/java/de/hysky/skyblocker/compatibility/MixinPlugin.java
@@ -2,6 +2,7 @@ package de.hysky.skyblocker.compatibility;
 
 import net.fabricmc.loader.api.FabricLoader;
 import org.objectweb.asm.tree.ClassNode;
+import org.spongepowered.asm.mixin.MixinEnvironment;
 import org.spongepowered.asm.mixin.extensibility.IMixinConfigPlugin;
 import org.spongepowered.asm.mixin.extensibility.IMixinInfo;
 
@@ -10,6 +11,7 @@ import java.util.Set;
 
 public class MixinPlugin implements IMixinConfigPlugin {
 	private static final boolean OPTIFABRIC_LOADED = FabricLoader.getInstance().isModLoaded("optifabric");
+	private static final boolean FCWLS_LOADED = FabricLoader.getInstance().isModLoaded("forcecloseworldloadingscreen");
 
 	@Override
 	public void onLoad(String mixinPackage) {
@@ -24,7 +26,16 @@ public class MixinPlugin implements IMixinConfigPlugin {
 	@Override
 	public boolean shouldApplyMixin(String targetClassName, String mixinClassName) {
 		//OptiFabric Compatibility
-        return !mixinClassName.endsWith("WorldRendererMixin") || !OPTIFABRIC_LOADED;
+		if (OPTIFABRIC_LOADED && mixinClassName.endsWith("WorldRendererMixin")) return false;
+
+		//Force Close World Screen Compatibility
+		if (FCWLS_LOADED && mixinClassName.endsWith("MinecraftClientScreenMixin")) {
+			String mixinCodeSource = MixinEnvironment.class.getProtectionDomain().getCodeSource().getLocation().toString();
+
+			return !mixinCodeSource.contains("0.12.5+mixin.0.8.5");
+		}
+
+        return true;
     }
 
 	@Override
