+++ src/main/resources/fabric.mod.json
@@ -41,9 +41,6 @@
   "conflicts": {
     "immediatelyfast": "<=1.2.10+1.20.4"
   },
-  "breaks": {
-    "forcecloseworldloadingscreen": "<=2.2.0"
-  },
   "custom": {
     "modmenu": {
       "links": {
