+++ src/main/resources/skyblocker.mixins.json
@@ -25,6 +25,7 @@
     "LeverBlockMixin",
     "LivingEntityRendererMixin",
     "MinecraftClientMixin",
+    "MinecraftClientScreenMixin",
     "PlayerInventoryMixin",
     "PlayerListHudMixin",
     "PlayerSkinProviderMixin",
